requirements.txt
# RicohLibrary – Phase 1‑4 dependencies
# Locked to stable versions tested on Python 3.11
# Streamlit will be added in Phase 5.

PyMuPDF==1.25.3
chromadb==0.6.3
rank_bm25==0.2.2
langgraph==0.2.74
langchain-core==0.3.34
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.2
